+++ Models/Data/Gallery.cs
@@ -1,9 +1,9 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
 using System.ComponentModel.DataAnnotations;
 
-namespace PhotoGalleryAPI.Models
+namespace PhotoGalleryAPI.Models.Data
 {
-    public class Photo
+    public class Gallery
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Key]
@@ -12,16 +12,18 @@ namespace PhotoGalleryAPI.Models
         [StringLength(100)]
         public string Name { get; set; } = string.Empty;
 
-        [StringLength(100)]
-        public Gallery Gallery { get; set; }
+        public User User { get; set; }
 
-        public Guid GalleryId { get; set; }
+        public Guid UserId { get; set; }
+
+        [StringLength(100)]
+        public string ParentName { get; set; } = string.Empty;
 
         [StringLength(250)]
         public string Description { get; set; } = string.Empty;
 
-        public byte[] ImageData { get; set; }
+        public List<Photo> Photos { get; set; } = new List<Photo>();
 
-        public DateTime UploadedAt { get; set; }
+        public DateTime CreatedAt { get; set; }
     }
 }
