+++ Models/Data/Photo.cs
@@ -1,9 +1,9 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
 using System.ComponentModel.DataAnnotations;
 
-namespace PhotoGalleryAPI.Models
+namespace PhotoGalleryAPI.Models.Data
 {
-    public class Gallery
+    public class Photo
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Key]
@@ -12,18 +12,16 @@ namespace PhotoGalleryAPI.Models
         [StringLength(100)]
         public string Name { get; set; } = string.Empty;
 
-        public User User { get; set; }
-
-        public Guid UserId { get; set; }
-
         [StringLength(100)]
-        public string ParentName { get; set; } = string.Empty;
+        public Gallery Gallery { get; set; }
+
+        public Guid GalleryId { get; set; }
 
         [StringLength(250)]
         public string Description { get; set; } = string.Empty;
 
-        public List<Photo> Photos { get; set; } = new List<Photo>();
+        public byte[] ImageData { get; set; }
 
-        public DateTime CreatedAt { get; set; }
+        public DateTime UploadedAt { get; set; }
     }
 }
