+++ Models/Dto/UserDto.cs
@@ -1,4 +1,4 @@
-﻿namespace PhotoGalleryAPI.Models
+﻿namespace PhotoGalleryAPI.Models.Dto
 {
     public record UserDto(string Name, string Password);
 }
