+++ Services/IUserService.cs
@@ -1,0 +1,10 @@
+﻿namespace PhotoGalleryAPI.Services
+{
+
+    public interface IUserService
+    {
+        void CreatePasswordHash(string password, out byte[] passwordHash, out byte[] passwordSalt);
+        bool VerifyPasswordHash(string password, byte[] passwordHash, byte[] passwordSalt);
+        string CreateToken(string username);
+    }
+}
