+++ Services/UserService.cs
@@ -5,12 +5,6 @@ using System.Security.Cryptography;
 
 namespace PhotoGalleryAPI.Services
 {
-    public interface IUserService
-    {
-        void CreatePasswordHash(string password, out byte[] passwordHash, out byte[] passwordSalt);
-        bool VerifyPasswordHash(string password, byte[] passwordHash, byte[] passwordSalt);
-        string CreateToken(string username);
-    }
 
     public class UserService : IUserService
     {
